⚡ update: tests, Clear-AugerContext to check if file exists

+++ Auger/Public/Clear-AugerContext.ps1
@@ -2,7 +2,9 @@ function Clear-AugerContext {
     [CmdletBinding()]
     param ()
 
-    Remove-Item $Script:AugerContext.LogFile.FullName
+    if ($Script:AugerContext.LogFile.FullName -and (Test-Path $Script:AugerContext.LogFile.FullName)) {
+        Remove-Item $Script:AugerContext.LogFile.FullName
+    }
 
     $Script:AugerContext = [pscustomobject]@{
         Application     = $null
